bugs fixed & cron add Signed-off-by: jiangchao1987 <jiangchao1987@gmail.com>

+++ src/main/java/com/candou/ac/rom/clean/CleanIllegalApp.java
@@ -13,29 +13,13 @@ import com.candou.conf.Configure;
  * 验证应用实际大小与网站记录的差值是否小于5M
  */
 public class CleanIllegalApp {
-//    private static final String path = "/data2/android/rom/roms";
-//	private static final String BASE_PATH = "C:/androidrom";
 	private static final String BASE_PATH = Configure.getProperty("rom_basepath");
     private static Logger logger = Logger.getLogger(CleanIllegalApp.class);
 
     public static void main(String[] args) {
-        /*File folder = new File(path);
-        logger.info("total find " + folder.listFiles().length + " files");
-        
-        File[] files = folder.listFiles();
-        for (File file : files) {
-            if (file.length() < 10 * 1024 * 1000) {
-                logger.info("illegal file [" + file.getName() + "]");
-                if (file.delete()) {
-                    // 将tb_app的filename更新为null
-                    RomAppDao.updateFileName(file.getName());
-                }
-            }
-        }*/
-    	
     	List<RomApp> apps = findApps();
     	for (RomApp app : apps) {
-    		if (!isValid(app)) {
+    		if (app.getSize() != 0.0f && !isValid(app)) {	//如果size为0说明这个字段没有抓取到
     			// delete && update filename to null
     			File illegalRom = new File(BASE_PATH + app.getFileName());
     			if (illegalRom.exists() && illegalRom.delete()) {
